requirements.txt
certifi==2021.5.30
numpy==1.19.5
pandas==1.1.5
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.4
schedule==1.1.0
six==1.16.0
wincertstore==0.2
